move drawer in separate component
