+++ src/components/drawer/drawer.module.css
@@ -1,0 +1,39 @@
+.drawer {
+	min-width: 300px;
+	display: flex;
+	flex-direction: column;
+}
+
+.header {
+	display: flex;
+	padding-bottom: 20px;
+	height: 40px;
+}
+
+.headerTitle {
+	flex-grow: 1;
+	font-size: 20px;
+	font-weight: bold;
+	line-height: 40px;
+}
+
+.closeButton {
+	flex-grow: 0;
+	width: 14px;
+	padding: 13px;
+	cursor: pointer;
+}
+
+.closeIcon {
+    width: 100%;
+    display: block;
+}
+
+.body {
+	display: flex;
+	flex-direction: column;
+	gap: 10px;
+	overflow-y: auto;
+	overflow-x: hidden;
+	flex-grow: 1;
+}
