+++ src/components/drawer/index.ts
@@ -1,0 +1,1 @@
+export { Drawer } from "./drawer";
